requirements.txt
icrawler>=0.6.6
playwright>=1.40.0
playwright-stealth>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
numpy>=1.24.0
